Roaster and dashboard chart

+++ app/Http/Controllers/Admin/OfficeHeaderController.php
@@ -5,16 +5,28 @@ namespace App\Http\Controllers\Admin;
 use App\Http\Controllers\Controller;
 use App\Http\Requests\OfficeHeader\UpdateOfficeHeaderRequest;
 use App\Models\OfficeHeader;
+use Illuminate\Support\Facades\Gate;
 
 class OfficeHeaderController extends Controller
 {
     public function edit(OfficeHeader $officeHeader)
     {
+        abort_if(
+            Gate::denies('officeHeader_edit'),
+            403,
+            'You are not allowed to access this resource'
+        );
         return view('admin.setting.officeSetting.edit', compact('officeHeader'));
     }
 
     public function update(UpdateOfficeHeaderRequest $request, OfficeHeader $officeHeader)
     {
+
+        abort_if(
+            Gate::denies('officeHeader_edit'),
+            403,
+            'You are not allowed to access this resource'
+        );
         $officeHeader->update($request->validated());
         toast('सफलतापूर्वक अद्यावधिक गरियो', 'success');
 
@@ -23,6 +35,12 @@ class OfficeHeaderController extends Controller
 
     public function destroy(OfficeHeader $officeHeader)
     {
+
+        abort_if(
+            Gate::denies('officeHeader_delete'),
+            403,
+            'You are not allowed to access this resource'
+        );
         $officeHeader->delete();
         toast('सफलतापूर्वक मेटियो', 'success');
